Moved scraper up a level
Use common module and added pydantic for settings
