+++ scraper/requirements.txt
@@ -3,3 +3,5 @@ requests
 structlog
 azure-cosmos
 dapr
+pydantic
+pydantic[dotenv]
